@@ -1,379 +1,433 @@
 import { IMXIcon } from "./common/icon";
+import brickWallIcon from "./svgs/brick-wall.svg";
+import brushIcon from "./svgs/brush.svg";
+import bucketIcon from "./svgs/bucket.svg";
+import buildingIcon from "./svgs/building.svg";
+import bulldozerIcon from "./svgs/bulldozer.svg";
+import cementMixerIcon from "./svgs/cement-mixer.svg";
+import constructionConeIcon from "./svgs/construction-cone.svg";
+import constructionProjectIcon from "./svgs/construction-project.svg";
+import constructionTrolleyIcon from "./svgs/construction-trolley.svg";
+import cordlessScrewdriverIcon from "./svgs/cordless-screwdriver.svg";
+import countryHomeIcon from "./svgs/country-home.svg";
+import craneIcon from "./svgs/crane.svg";
+import dumpTruckIcon from "./svgs/dump-truck.svg";
+import estateIcon from "./svgs/estate.svg";
+import forkliftTruckIcon from "./svgs/forklift-truck.svg";
+import foundationIcon from "./svgs/foundation.svg";
+import fountainIcon from "./svgs/fountain.svg";
+import hammerIcon from "./svgs/hammer.svg";
+import impactDrillIcon from "./svgs/impact-drill.svg";
+import ladderIcon from "./svgs/ladder.svg";
+import maintenanceIcon from "./svgs/maintenance.svg";
+import oilPlatformIcon from "./svgs/oil-platform.svg";
+import paintRollerIcon from "./svgs/paint-roller.svg";
+import pipesIcon from "./svgs/pipes.svg";
+import pipingSystemIcon from "./svgs/piping-system.svg";
+import pliersIcon from "./svgs/pliers.svg";
+import pneumaticHammerDrillIcon from "./svgs/pneumatic-hammer-drill.svg";
+import radiatorIcon from "./svgs/radiator.svg";
+import restrictedAreaIcon from "./svgs/restricted-area.svg";
+import roadRollerIcon from "./svgs/road-roller.svg";
+import rullerIcon from "./svgs/rulers.svg";
+import safetyVestIcon from "./svgs/safety-vest.svg";
+import sawIcon from "./svgs/saw.svg";
+import scissorLiftIcon from "./svgs/scissor-lift.svg";
+import screwdriverIcon from "./svgs/screwdriver.svg";
+import shovelIcon from "./svgs/shovel.svg";
+import skyscrapersIcon from "./svgs/skyscrapers.svg";
+import solarPanelIcon from "./svgs/solar-panel.svg";
+import solderingIronIcon from "./svgs/soldering-iron.svg";
+import stairsIcon from "./svgs/stairs.svg";
+import stopHandIcon from "./svgs/stop-hand.svg";
+import tableSawBladeIcon from "./svgs/table-saw-blade.svg";
+import tapWaterIcon from "./svgs/tap-water.svg";
+import tapelineIcon from "./svgs/tapeline.svg";
+import towTruckIcon from "./svgs/tow-truck.svg";
+import towerRailIcon from "./svgs/towel-rail.svg";
+import tractorIcon from "./svgs/tractor.svg";
+import trowelIcon from "./svgs/trowel.svg";
+import waterBoilerIcon from "./svgs/water-boiler.svg";
+import waterHeaterIcon from "./svgs/water-heater.svg";
+import waterTapIcon from "./svgs/water-tap.svg";
+import workerIcon from "./svgs/worker.svg";
+import workersIcon from "./svgs/workers.svg";
+import wrenchIcon from "./svgs/wrench.svg";
 
 export const brickWall: IMXIcon = {
   name: "brick-wall",
-  value: require("./svgs/brick-wall.svg"),
+  value: brickWallIcon,
   categories: ["construction"],
   keywords: ["brick", "wall", "block", "pattern", "architecture", "stone", "building", "construction", "house"],
 };
 
 export const brush: IMXIcon = {
   name: "brush",
-  value: require("./svgs/brush.svg"),
+  value: brushIcon,
   categories: ["construction"],
   keywords: ["brush", "decoration", "interior", "paint", "painting", "roller", "wall"],
 };
 
 export const bucket: IMXIcon = {
   name: "bucket",
-  value: require("./svgs/bucket.svg"),
+  value: bucketIcon,
   categories: ["construction"],
   keywords: ["bucket", "decoration", "interior", "paint", "water", "house", "wall", "construction"],
 };
 
 export const building: IMXIcon = {
   name: "building",
-  value: require("./svgs/building.svg"),
+  value: buildingIcon,
   categories: ["construction"],
   keywords: ["building", "home", "house", "apartments", "architecture", "structure"],
 };
 
 export const bulldozer: IMXIcon = {
   name: "bulldozer",
-  value: require("./svgs/bulldozer.svg"),
+  value: bulldozerIcon,
   categories: ["construction"],
   keywords: ["bulldozer", "digging", "excavator", "machine"],
 };
 
 export const cementMixer: IMXIcon = {
   name: "cement-mixer",
-  value: require("./svgs/cement-mixer.svg"),
+  value: cementMixerIcon,
   categories: ["construction"],
   keywords: ["cement", "mixer", "concrete", "construction", "machine", "truck"],
 };
 
 export const constructionCone: IMXIcon = {
   name: "construction-cone",
-  value: require("./svgs/construction-cone.svg"),
+  value: constructionConeIcon,
   categories: ["construction"],
   keywords: ["construction", "cone", "barrier", "caution", "road", "safety"],
 };
 
 export const constructionProject: IMXIcon = {
   name: "construction-project",
-  value: require("./svgs/construction-project.svg"),
+  value: constructionProjectIcon,
   categories: ["construction"],
   keywords: ["construction", "project", "blueprint", "plan", "architecture"],
 };
 
 export const constructionTrolley: IMXIcon = {
   name: "construction-trolley",
-  value: require("./svgs/construction-trolley.svg"),
+  value: constructionTrolleyIcon,
   categories: ["construction"],
   keywords: ["construction", "trolley", "carrier", "cart", "cement", "holder"],
 };
 
 export const cordlessScrewdriver: IMXIcon = {
   name: "cordless-screwdriver",
-  value: require("./svgs/cordless-screwdriver.svg"),
+  value: cordlessScrewdriverIcon,
   categories: ["construction"],
   keywords: ["cordless", "screwdriver", "drill", "driver", "electric", "power"],
 };
 
 export const countryHome: IMXIcon = {
   name: "country-home",
-  value: require("./svgs/country-home.svg"),
+  value: countryHomeIcon,
   categories: ["construction"],
   keywords: ["country", "home", "house", "real estate"],
 };
 
 export const crane: IMXIcon = {
   name: "crane",
-  value: require("./svgs/crane.svg"),
+  value: craneIcon,
   categories: ["construction"],
   keywords: ["crane", "construction", "hook", "machinery", "tower"],
 };
 
 export const dumpTruck: IMXIcon = {
   name: "dump-truck",
-  value: require("./svgs/dump-truck.svg"),
+  value: dumpTruckIcon,
   categories: ["construction"],
   keywords: ["dump", "truck", "dumper", "garbage", "trash", "waste", "vehicle", "transport"],
 };
 
 export const estate: IMXIcon = {
   name: "estate",
-  value: require("./svgs/estate.svg"),
+  value: estateIcon,
   categories: ["construction"],
   keywords: ["estate", "building", "home", "house", "architecture"],
 };
 
 export const forkliftTruck: IMXIcon = {
   name: "forklift-truck",
-  value: require("./svgs/forklift-truck.svg"),
+  value: forkliftTruckIcon,
   categories: ["construction"],
   keywords: ["forklift", "truck", "warehouse", "industrial", "transport", "lift", "cargo", "logistics", "loader", "machinery"],
 };
 
 export const foundation: IMXIcon = {
   name: "foundation",
-  value: require("./svgs/foundation.svg"),
+  value: foundationIcon,
   categories: ["construction"],
   keywords: ["foundation", "wall", "architecture", "construction"],
 };
 
 export const fountain: IMXIcon = {
   name: "fountain",
-  value: require("./svgs/fountain.svg"),
+  value: fountainIcon,
   categories: ["construction"],
   keywords: ["fountain", "garden", "splash", "park", "water"],
 };
 
 export const hammer: IMXIcon = {
   name: "hammer",
-  value: require("./svgs/hammer.svg"),
+  value: hammerIcon,
   categories: ["construction"],
   keywords: ["hammer", "construction", "equipment", "hardware", "repair", "tool", "work", "tools"],
 };
 
 export const impactDrill: IMXIcon = {
   name: "impact-drill",
-  value: require("./svgs/impact-drill.svg"),
+  value: impactDrillIcon,
   categories: ["construction"],
   keywords: ["impact drill", "power", "tools", "driver", "construction", "tool", "work", "repair", "drill"],
 };
 
 export const ladder: IMXIcon = {
   name: "ladder",
-  value: require("./svgs/ladder.svg"),
+  value: ladderIcon,
   categories: ["construction"],
   keywords: ["ladder", "construction", "building", "climb", "height", "tool", "work", "tools"],
 };
 
 export const maintenance: IMXIcon = {
   name: "maintenance",
-  value: require("./svgs/maintenance.svg"),
+  value: maintenanceIcon,
   categories: ["construction"],
   keywords: ["maintenance", "construction", "tools", "hardware", "equipment", "wrench", "screwdriver", "tool kit", "tool box"],
 };
 
 export const oilPlatform: IMXIcon = {
   name: "oil-platform",
-  value: require("./svgs/oil-platform.svg"),
+  value: oilPlatformIcon,
   categories: ["construction"],
   keywords: ["oil platform", "industry", "drilling", "platform", "oil", "rig", "offshore"],
 };
 
 export const paintRoller: IMXIcon = {
   name: "paint-roller",
-  value: require("./svgs/paint-roller.svg"),
+  value: paintRollerIcon,
   categories: ["construction"],
   keywords: ["paint roller", "equipment", "house construction", "paint", "roller", "wall paint", "tools", "tool"],
 };
 
 export const pipes: IMXIcon = {
   name: "pipes",
-  value: require("./svgs/pipes.svg"),
+  value: pipesIcon,
   categories: ["construction"],
   keywords: ["pipes", "plumbing", "pipe", "drain", "water"],
 };
 
 export const pipingSystem: IMXIcon = {
   name: "piping-system",
-  value: require("./svgs/piping-system.svg"),
+  value: pipingSystemIcon,
   categories: ["construction"],
   keywords: ["piping system", "piping", "water", "drain", "water", "industrial"],
 };
 
 export const pliers: IMXIcon = {
   name: "pliers",
-  value: require("./svgs/pliers.svg"),
+  value: pliersIcon,
   categories: ["construction"],
   keywords: ["pliers", "equipment", "tools", "tool", "repairs", "hardware"],
 };
 
 export const pneumaticHammerDrill: IMXIcon = {
   name: "pneumatic-hammer-drill",
-  value: require("./svgs/pneumatic-hammer-drill.svg"),
+  value: pneumaticHammerDrillIcon,
   categories: ["construction"],
   keywords: ["pneumatic hammer drill", "drill", "hammer", "hydraulic", "jack", "kit", "pneumatic", "asphalt"],
 };
 
 export const radiator: IMXIcon = {
   name: "radiator",
-  value: require("./svgs/radiator.svg"),
+  value: radiatorIcon,
   categories: ["construction"],
   keywords: ["radiator", "heat", "warm", "interior", "heater", "heating"],
 };
 
 export const restrictedArea: IMXIcon = {
   name: "restricted-area",
-  value: require("./svgs/restricted-area.svg"),
+  value: restrictedAreaIcon,
   categories: ["construction"],
   keywords: ["restricted area", "zone", "no entry", "warning sign", "caution zone", "area", "military", "restricted"],
 };
 
 export const roadRoller: IMXIcon = {
   name: "road-roller",
-  value: require("./svgs/road-roller.svg"),
+  value: roadRollerIcon,
   categories: ["construction"],
   keywords: ["road roller", "construction", "road", "roller", "vehicle", "roadwork"],
 };
 
 export const rulers: IMXIcon = {
   name: "rulers",
-  value: require("./svgs/rulers.svg"),
+  value: rullerIcon,
   categories: ["construction"],
   keywords: ["rulers", "measure", "scale", "measurement", "centimeter", "tools"],
 };
 
 export const safetyVest: IMXIcon = {
   name: "safety-vest",
-  value: require("./svgs/safety-vest.svg"),
+  value: safetyVestIcon,
   categories: ["construction"],
   keywords: ["safety vest", "construction", "engineer", "jacket", "protect", "safety"],
 };
 
 export const saw: IMXIcon = {
   name: "saw",
-  value: require("./svgs/saw.svg"),
+  value: sawIcon,
   categories: ["construction"],
   keywords: ["saw", "hand", "tools", "woodworking", "equipment", "tool"],
 };
 
 export const scissorLift: IMXIcon = {
   name: "scissor-lift",
-  value: require("./svgs/scissor-lift.svg"),
+  value: scissorLiftIcon,
   categories: ["construction"],
   keywords: ["scissor lift", "lift", "scaffolding", "scissor", "hydraulic", "construction", "high", "factory", "industry"],
 };
 
 export const screwdriver: IMXIcon = {
   name: "screwdriver",
-  value: require("./svgs/screwdriver.svg"),
+  value: screwdriverIcon,
   categories: ["construction"],
   keywords: ["screwdriver", "tools", "tool", "hardware", "screw", "fixer", "repair"],
 };
 
 export const shovel: IMXIcon = {
   name: "shovel",
-  value: require("./svgs/shovel.svg"),
+  value: shovelIcon,
   categories: ["construction"],
   keywords: ["shovel", "digging", "tool", "spade", "equipment"],
 };
 
 export const skyscrapers: IMXIcon = {
   name: "skyscrapers",
-  value: require("./svgs/skyscrapers.svg"),
+  value: skyscrapersIcon,
   categories: ["construction"],
   keywords: ["skyscrapers", "buildings", "city", "cityscape", "downtown", "skyline", "architecture"],
 };
 
 export const solarPanel: IMXIcon = {
   name: "solar-panel",
-  value: require("./svgs/solar-panel.svg"),
+  value: solarPanelIcon,
   categories: ["construction"],
   keywords: ["solar panel", "alternative", "energy", "electricity", "panel", "solar"],
 };
 
 export const solderingIron: IMXIcon = {
   name: "soldering-iron",
-  value: require("./svgs/soldering-iron.svg"),
+  value: solderingIronIcon,
   categories: ["construction"],
   keywords: ["soldering iron", "soldering", "iron", "tool", "copper", "manufactoring", "electronic"],
 };
 
 export const stairs: IMXIcon = {
   name: "stairs",
-  value: require("./svgs/stairs.svg"),
+  value: stairsIcon,
   categories: ["construction"],
   keywords: ["stairs", "building", "climb", "height"],
 };
 
 export const stopHand: IMXIcon = {
   name: "stop-hand",
-  value: require("./svgs/stop-hand.svg"),
+  value: stopHandIcon,
   categories: ["construction"],
   keywords: ["stop hand", "stop", "sign", "halt", "adblock"],
 };
 
 export const tableSawBlade: IMXIcon = {
   name: "table-saw-blade",
-  value: require("./svgs/table-saw-blade.svg"),
+  value: tableSawBladeIcon,
   categories: ["construction"],
   keywords: ["table saw", "blade", "cut", "saw", "table"],
 };
 
 export const tapWater: IMXIcon = {
   name: "tap-water",
-  value: require("./svgs/tap-water.svg"),
+  value: tapWaterIcon,
   categories: ["construction"],
   keywords: ["tap water", "tap", "water", "faucet", "drip", "leak"],
 };
 
 export const tapeline: IMXIcon = {
   name: "tapeline",
-  value: require("./svgs/tapeline.svg"),
+  value: tapelineIcon,
   categories: ["construction"],
   keywords: ["tapeline", "centimeter", "inches tape", "measuring tape", "meter stick", "inches", "measure", "metre", "ruler"],
 };
 
 export const towTruck: IMXIcon = {
   name: "tow-truck",
-  value: require("./svgs/tow-truck.svg"),
+  value: towTruckIcon,
   categories: ["construction"],
   keywords: ["tow truck", "repair", "service", "tow", "truck", "towing", "vehicle", "transport"],
 };
 
 export const towelRail: IMXIcon = {
   name: "towel-rail",
-  value: require("./svgs/towel-rail.svg"),
+  value: towerRailIcon,
   categories: ["construction"],
   keywords: ["towel rail", "warmer", "heated", "towel", "rail", "bathroom", "interior", "rack"],
 };
 
 export const tractor: IMXIcon = {
   name: "tractor",
-  value: require("./svgs/tractor.svg"),
+  value: tractorIcon,
   categories: ["construction"],
   keywords: ["tractor", "vehicle", "machinery"],
 };
 
 export const trowel: IMXIcon = {
   name: "trowel",
-  value: require("./svgs/trowel.svg"),
+  value: trowelIcon,
   categories: ["construction"],
   keywords: ["trowel", "cement", "equipment", "tool", "renovation", "maintenance"],
 };
 
 export const waterBoiler: IMXIcon = {
   name: "water-boiler",
-  value: require("./svgs/water-boiler.svg"),
+  value: waterBoilerIcon,
   categories: ["construction"],
   keywords: ["water boiler", "boiler", "heater", "water", "heating", "appliance"],
 };
 
 export const waterHeater: IMXIcon = {
   name: "water-heater",
-  value: require("./svgs/water-heater.svg"),
+  value: waterHeaterIcon,
   categories: ["construction"],
   keywords: ["water heater", "heater", "water", "tank", "appliance", "electrical", "heating"],
 };
 
 export const waterTap: IMXIcon = {
   name: "water-tap",
-  value: require("./svgs/water-tap.svg"),
+  value: waterTapIcon,
   categories: ["construction"],
   keywords: ["water tap", "tap", "faucet", "water", "plumbing", "valve"],
 };
 
 export const worker: IMXIcon = {
   name: "worker",
-  value: require("./svgs/worker.svg"),
+  value: workerIcon,
   categories: ["construction"],
   keywords: ["worker", "builder", "engineer", "man", "labor"],
 };
 
 export const workers: IMXIcon = {
   name: "workers",
-  value: require("./svgs/workers.svg"),
+  value: workersIcon,
   categories: ["construction"],
   keywords: ["workers", "crew", "builders", "engineers", "men", "constructors"],
 };
 
 export const wrench: IMXIcon = {
   name: "wrench",
-  value: require("./svgs/wrench.svg"),
+  value: wrenchIcon,
   categories: ["construction"],
   keywords: ["wrench", "construction", "equipment", "hardware", "metal", "tool"],
 };
